♻️ refactoring : selectStack hook -> functions로 변경
