+++ client/src/components/mainPage/StackAtom.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import Stack from './Stack';
 import { stackFamily } from '../../states/mainPage';
-import { useRecoilValue } from 'recoil';
-import useStackSelect from '../../hooks/useStackSelect';
+import { useRecoilCallback, useRecoilValue } from 'recoil';
+import selectStack from '../../functions/selectStack';
 
 interface StackAtomProps {
   value: string;
@@ -11,9 +11,9 @@ interface StackAtomProps {
 const StackAtom = ({ value }: StackAtomProps) => {
   const stack = useRecoilValue(stackFamily(value));
 
-  const selectStack = useStackSelect(value);
+  const selectHandler = useRecoilCallback(selectStack(value), []);
 
-  return <Stack onClick={selectStack} value={value} selected={stack.selected} />;
+  return <Stack onClick={selectHandler} value={value} selected={stack.selected} />;
 };
 
 export default StackAtom;
