--- client/src/hooks/useStackSelect.ts
@@ -1,29 +1,0 @@
-import { stackFamily, stackState } from './../states/mainPage';
-import { useRecoilCallback } from 'recoil';
-
-const useStackSelect = (value: string) => {
-  const selectStack = useRecoilCallback(
-    ({ snapshot, set }) =>
-      () => {
-        const stacks = snapshot.getLoadable(stackState).getValue();
-        const currentStack = snapshot.getLoadable(stackFamily(value)).getValue();
-        const currentStackAtom = stackFamily(value);
-
-        if (!currentStack.selected) {
-          set(stackState, [...stacks, value]);
-        } else {
-          set(stackState, [...stacks.filter((stack) => stack !== value)]);
-        }
-
-        set(currentStackAtom, {
-          ...currentStack,
-          selected: !currentStack.selected,
-        });
-      },
-    [],
-  );
-
-  return selectStack;
-};
-
-export default useStackSelect;
